refactor(testkit): host per-service SDK client factories behind a feature
Moves the ~20 per-service AWS SDK client factories (sqs_client,
sns_client, s3_client, ...) and the `aws_cli` CLI wrapper from the e2e
and conformance helper modules into fakecloud_testkit itself, gated
behind a new optional `sdk-clients` feature. Both test crates now just
re-export `fakecloud_testkit::TestServer` and enable the feature in
their dev-deps.

fakecloud-tfacc keeps the feature off so its lifecycle-only wrapper
doesn't pay the aws-sdk-* compile cost.

Single source of truth for client construction quirks — notably
`force_path_style(true)` for S3 against fakecloud.
