+++ crates/fakecloud-conformance/Cargo.toml
@@ -21,7 +21,7 @@ regex = "1"
 
 [dev-dependencies]
 fakecloud-conformance-macros = { path = "../fakecloud-conformance-macros" }
-fakecloud-testkit = { path = "../fakecloud-testkit" }
+fakecloud-testkit = { path = "../fakecloud-testkit", features = ["sdk-clients"] }
 aws-config = "1"
 aws-sdk-sqs = "1"
 aws-sdk-sns = "1"
