+++ crates/fakecloud-e2e/Cargo.toml
@@ -45,6 +45,6 @@ base64 = { workspace = true }
 flate2 = { workspace = true }
 zip = { workspace = true }
 fakecloud-sdk = { path = "../fakecloud-sdk" }
-fakecloud-testkit = { path = "../fakecloud-testkit" }
+fakecloud-testkit = { path = "../fakecloud-testkit", features = ["sdk-clients"] }
 tempfile = { workspace = true }
 sha2 = { workspace = true }
