+++ crates/fakecloud-testkit/src/lib.rs
@@ -5,8 +5,10 @@
 //! rolling its own process-lifecycle code.
 //!
 //! Scope is intentionally narrow: spawn / endpoint / `SdkConfig` / cleanup.
-//! Per-service SDK client factories live in each consumer so this crate
-//! doesn't drag in every `aws-sdk-*` dependency.
+//!
+//! Per-service AWS SDK client factories are available behind the optional
+//! `sdk-clients` feature. Lifecycle-only consumers (e.g. `fakecloud-tfacc`)
+//! can keep the feature off and avoid compiling the `aws-sdk-*` crates.
 
 use std::net::TcpListener;
 use std::path::{Path, PathBuf};
@@ -180,6 +182,25 @@ impl TestServer {
             ))
             .load()
             .await
+    }
+
+    /// Invoke the system `aws` CLI against this server with `test`
+    /// credentials and `us-east-1` region. Returns the raw `Output` wrapped
+    /// in `CliOutput` so callers can use the `stdout_json` / `stdout_text`
+    /// helpers.
+    pub async fn aws_cli(&self, args: &[&str]) -> CliOutput {
+        let output = Command::new("aws")
+            .args(args)
+            .arg("--endpoint-url")
+            .arg(self.endpoint())
+            .arg("--region")
+            .arg("us-east-1")
+            .env("AWS_ACCESS_KEY_ID", "test")
+            .env("AWS_SECRET_ACCESS_KEY", "test")
+            .env("AWS_DEFAULT_REGION", "us-east-1")
+            .output()
+            .expect("failed to run aws cli");
+        CliOutput(output)
     }
 }
 
@@ -404,4 +425,105 @@ async fn wait_for_port(child: &mut Child, port: u16) -> bool {
         tokio::time::sleep(Duration::from_millis(100)).await;
     }
     false
+}
+
+#[cfg(feature = "sdk-clients")]
+impl TestServer {
+    pub async fn sqs_client(&self) -> aws_sdk_sqs::Client {
+        aws_sdk_sqs::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn sns_client(&self) -> aws_sdk_sns::Client {
+        aws_sdk_sns::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn eventbridge_client(&self) -> aws_sdk_eventbridge::Client {
+        aws_sdk_eventbridge::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn iam_client(&self) -> aws_sdk_iam::Client {
+        aws_sdk_iam::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn sts_client(&self) -> aws_sdk_sts::Client {
+        aws_sdk_sts::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn ssm_client(&self) -> aws_sdk_ssm::Client {
+        aws_sdk_ssm::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn dynamodb_client(&self) -> aws_sdk_dynamodb::Client {
+        aws_sdk_dynamodb::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn lambda_client(&self) -> aws_sdk_lambda::Client {
+        aws_sdk_lambda::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn secretsmanager_client(&self) -> aws_sdk_secretsmanager::Client {
+        aws_sdk_secretsmanager::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn logs_client(&self) -> aws_sdk_cloudwatchlogs::Client {
+        aws_sdk_cloudwatchlogs::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn kms_client(&self) -> aws_sdk_kms::Client {
+        aws_sdk_kms::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn kinesis_client(&self) -> aws_sdk_kinesis::Client {
+        aws_sdk_kinesis::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn rds_client(&self) -> aws_sdk_rds::Client {
+        aws_sdk_rds::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn elasticache_client(&self) -> aws_sdk_elasticache::Client {
+        aws_sdk_elasticache::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn cloudformation_client(&self) -> aws_sdk_cloudformation::Client {
+        aws_sdk_cloudformation::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn ses_client(&self) -> aws_sdk_ses::Client {
+        aws_sdk_ses::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn sesv2_client(&self) -> aws_sdk_sesv2::Client {
+        aws_sdk_sesv2::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn cognito_client(&self) -> aws_sdk_cognitoidentityprovider::Client {
+        aws_sdk_cognitoidentityprovider::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn sfn_client(&self) -> aws_sdk_sfn::Client {
+        aws_sdk_sfn::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn apigatewayv2_client(&self) -> aws_sdk_apigatewayv2::Client {
+        aws_sdk_apigatewayv2::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn bedrock_client(&self) -> aws_sdk_bedrock::Client {
+        aws_sdk_bedrock::Client::new(&self.aws_config().await)
+    }
+
+    pub async fn bedrock_runtime_client(&self) -> aws_sdk_bedrockruntime::Client {
+        aws_sdk_bedrockruntime::Client::new(&self.aws_config().await)
+    }
+
+    /// S3 always uses path-style addressing against fakecloud so bucket
+    /// names don't have to be DNS-legal subdomains of `localhost`.
+    pub async fn s3_client(&self) -> aws_sdk_s3::Client {
+        let config = self.aws_config().await;
+        let s3_config = aws_sdk_s3::config::Builder::from(&config)
+            .force_path_style(true)
+            .build();
+        aws_sdk_s3::Client::from_conf(s3_config)
+    }
 }
